Xamarin: fix enum backing types
Objective Sharpie generates a fake backing type for enums backed by
native types in Objective-C (NSInteger, NSUInteger). This is a bug
in Objective Sharpie.

We need to change these to real C# types that are allowed to back
enums.

In Xamarin bindings, all enums should be 64 bits when backed by
native types in Objective-C.