Xamarin: audit MethodToProperty [Verify] warnings
When Objective Sharpie is not 100% sure of a binding decision it makes,
due to a lack of metadata in the Objective-C API it’s binding, it
applies a [Verify] attribute with a hint for cross referencing in
documentation.

[Verify (MethodToProperty)] means that Objective Sharpie bound an
Objective-C method as a C# property to make the C# API more friendly.
Most of the time this is correct, but it’s always best to double check
to see if a property really makes the most sense for the fluidity
of the binding API.

In this case, properties make the most sense for the methods
Objective Sharpie encountered that were “property-esque.”