Xamarin: audit [Verify (StronlgyTypedNSArray)]
In many cases we can strongly-type an array, but Objective Sharpie
may not know how since the Objective-C typing for the array is
weak (NSArray). Documentation and context helps us to infer the
contents of the array and can allow us to define a strong binding.

In some cases NSObject[] (weak typing) is still necessary. For instance,
many methods here take “a track or URI”. Because SPTTrack and NSUrl
are disjoint types, NSObject is appropriate.